Try-claim decorator converted to an object (#55)
* change the try-claim decorator to an object

* claim decorator takes functions it needs, no longer relies on power being decorated

Previously the claim decorator assumed that the power_on function was decorated with the same decorator. If power on was requested, the power on function would do the claiming, and then power on the system. That is kind of confusing, so now the decorator will call claim first, and then optionally power on.

Also takes the power on and claim functions, rather than calling the ones on the wrapper itself.

* formatting

* move decorator object to helper file, rename and make functions better, decorate docking functions

* formatting

* remove mistakenly merged robot command data and state

* decorate arm commands

* Fixed bugs with decorated functions

* black

---------

Co-authored-by: Andrew Messing <129519955+amessing-bdai@users.noreply.github.com>
Co-authored-by: Andrew Messing <amessing@theaiinstitute.com>