[SW-743] Add mutate_world_objects functionality to spot_wrapper (#87)
This allows the user to mutate Spot's world objects. Specifically, it can be used to create/delete no-go regions for Spot which are implemented as new world objects. 

I have been testing this in conjunction with a [follow-up PR](https://github.com/bdaiinstitute/spot_ros2/pull/271) to integrate this into Spot ROS 2. With this implementation, Spot is successfully able to receive and respond to mutate world object requests.