[SW-1245] Custom Gripper Calibration for Spot (#131)
# Change Overview
Related PR for ROS2 Functionality: https://github.com/bdaiinstitute/spot_ros2/pull/455#issue-2474033396

Add a custom automatic calibration routine to allow Spot to calibrate it's own hand cameras. 

# Testing Done
Ran the entire calibration, end-to-end, several times on a few different Spots and calibration boards in Spot lab. Also, ran different calibrations on the same collected datasets.

Note: Open3d is added as a dep so that the republisher in the related PR that shares the same requirements.txt is able to do point cloud to image conversions