[MAPLE-663] Update gripper_angle_open function (#142)
This is a slight update to the `gripper_angle_open` function that already existed in `spot_arm.py`. Simply Adds a flag to allow running this function without forcing a power on and stand of the robot. 
Also a slight update to `block_until_gripper_command_completes` as all gripper commands were technically failing because the feedback structure perhaps changed recently