[N/A] Flag to disable obstacle avoidance in trajectory_cmd (#163)
Adds a flag to disable obstacle avoidance when executing a trajectory command.
Despite `mobility_params` being an argument to `trajectory_cmd` in order set this param higher up in the pipeline, we would need to import `spot_command_pb2.MobilityParams` higher up, which is a can of worms I would rather not do when it can be solved with a simple boolean here. Open to discussion though if anyone feels strongly.